fix a bug on marketplace
