+++ src/constants/plugin-tags.ts
@@ -1,0 +1,16 @@
+import type { PluginTag } from './plugins'
+
+export const allTags: PluginTag[] = [
+  'New & Upcoming',
+  'Theme',
+  'Memberships',
+  'Governance',
+  'Crowdfunding',
+  'Messaging',
+  'Widgets',
+  'Commerce & Business',
+  'Finance & Payments',
+  'Analytics',
+  'Community',
+  'Utility',
+]
