+++ src/constants/plugins.ts
@@ -48,21 +48,6 @@ export type PluginTag =
   | 'Utility'
   | 'Community'
 
-export const allTags: PluginTag[] = [
-  'New & Upcoming',
-  'Theme',
-  'Memberships',
-  'Governance',
-  'Crowdfunding',
-  'Messaging',
-  'Widgets',
-  'Commerce & Business',
-  'Finance & Payments',
-  'Analytics',
-  'Community',
-  'Utility',
-]
-
 export type InstallablePlugins = {
   id: string
   tag: PluginTag
