+++ src/plugins/marketplace/admin.astro
@@ -9,7 +9,7 @@ import Heading from './assets/Heading.jpg'
 import { createGetPluginMeta } from './utils/createGetPluginMeta'
 import PluginsWrapper from './components/PluginsWrapper.astro'
 import Nav from './components/Nav.svelte'
-import { allTags } from '@constants/plugins'
+import { allTags } from '@constants/plugin-tags'
 import BannerPresetFeaturing from './components/BannerPresetFeaturing.astro'
 import ExternalToolWrapper from './components/ExternalToolWrapper.astro'
 import {
@@ -64,7 +64,7 @@ const externalTools: ExternalTool[] = [
   </section>
 
   <div class="mt-10">
-    <Nav client:load plugins={allPlugins} />
+    <Nav plugins={allPlugins} />
     <!-- New and Upcoming section -->
     {
       (
@@ -120,3 +120,4 @@ const externalTools: ExternalTool[] = [
     />
   </div>
 </div>
+@constants/plugin-tags
